external-match-client: Add order book depth helper
